refactor(auth config): define a new data model for auth config
